+++ crates/total-viewsheds/src/cpu/rotation.rs
@@ -11,7 +11,7 @@ use kernel::rotation::ANGLE_SHIFT;
     clippy::cast_precision_loss,
     reason = "so long as max_los^2 < 2^24, the following `as` conversions are entirely safe"
 )]
-pub fn generate_rotation(elevs: &[i16], angle: f32, max_los: usize) -> (Vec<i32>, Vec<i16>) {
+pub fn generate_rotation(elevs: &[i16], angle: f64, max_los: usize) -> (Vec<i64>, Vec<i16>) {
     let width = (max_los * 3) as isize;
     #[expect(clippy::integer_division, reason = "we don't need precision here")]
     {
@@ -31,27 +31,27 @@ pub fn generate_rotation(elevs: &[i16], angle: f32, max_los: usize) -> (Vec<i32>
     };
 
     let (sin, cos) = (
-        f32::sin((angle + ANGLE_SHIFT).to_radians()),
-        f32::cos((angle + ANGLE_SHIFT).to_radians()),
+        f64::sin((angle + f64::from(ANGLE_SHIFT)).to_radians()),
+        f64::cos((angle + f64::from(ANGLE_SHIFT)).to_radians()),
     );
 
-    let (x_center, y_center) = ((width - 1) as f32 / 2.0, (width - 1) as f32 / 2.0);
+    let (x_center, y_center) = ((width - 1) as f64 / 2.0, (width - 1) as f64 / 2.0);
 
-    let mut rotation: Vec<i32> = Vec::with_capacity(2 * max_los * max_los);
+    let mut rotation: Vec<i64> = Vec::with_capacity(2 * max_los * max_los);
 
     for x in (max_los as isize)..(max_los as isize) * 2 {
-        let x_sin = (x as f32 - x_center) * sin;
-        let x_cos = (x as f32 - x_center) * cos;
+        let x_sin = (x as f64 - x_center) * sin;
+        let x_cos = (x as f64 - x_center) * cos;
         for y in (max_los as isize)..width {
-            let y_sin = (y as f32 - y_center) * sin;
-            let y_cos = (y as f32 - y_center) * cos;
+            let y_sin = (y as f64 - y_center) * sin;
+            let y_cos = (y as f64 - y_center) * cos;
 
             let x_rot = (x_cos - y_sin + y_center).round() as isize;
             let y_rot = (y_cos + x_sin + x_center).round() as isize;
 
             let new_idx = x_rot.clamp(0, width - 1) * width + y_rot.clamp(0, width - 1);
 
-            rotation.push(new_idx as i32);
+            rotation.push(new_idx as i64);
         }
     }
 
@@ -65,7 +65,7 @@ pub fn generate_rotation(elevs: &[i16], angle: f32, max_los: usize) -> (Vec<i32>
     let elevations = rotation
         .iter()
         .map(|&idx| {
-            if idx < 0i32 {
+            if idx < 0i64 {
                 i16::MIN
             } else {
                 #[expect(
